refactor: move getAuthorNoteDefaultText from util.ts to process/prompt.ts
