+++ src/lib/SideBars/CharConfig.svelte
@@ -11,7 +11,8 @@
     import LoreBook from "./LoreBook/LoreBookSetting.svelte";
     import { alertTOS, showHypaV2Alert } from "../../ts/alert";
     import BarIcon from "./BarIcon.svelte";
-    import { getAuthorNoteDefaultText, selectMultipleFile, selectSingleFile } from "../../ts/util";
+    import { selectMultipleFile, selectSingleFile } from "../../ts/util";
+    import { getAuthorNoteDefaultText } from "src/ts/process/prompt"
     import { findCharacterbyId } from "../../ts/characters";
     import Help from "../Others/Help.svelte";
     import { exportChar } from "src/ts/characterCards";
