+++ src/ts/process/index.svelte.ts
@@ -9,7 +9,8 @@ import { ChatTokenizer, tokenize, tokenizeNum } from "../tokenizer";
 import { language } from "../../lang";
 import { alertError, alertToast } from "../alert";
 import { loadLoreBookV3Prompt } from "./lorebook.svelte";
-import { getAuthorNoteDefaultText, isLastCharPunctuation, trimUntilPunctuation, parseToggleSyntax } from "../util";
+import { isLastCharPunctuation, trimUntilPunctuation, parseToggleSyntax } from "../util";
+import { getAuthorNoteDefaultText } from "./prompt";
 import { findCharacterbyId } from "../characters";
 import { getPersonaPrompt } from "../persona";
 import { getUserName } from "../persona";
